Move slate into Ext like other extensions
